+++ archr/arsenal/qemu_tracer.py
@@ -13,7 +13,6 @@ import os
 l = logging.getLogger("archr.arsenal.qemu_tracer")
 
 from . import ContextBow
-from ..arsenal import Flight
 
 class QemuTraceResult:
     # results
@@ -74,17 +73,15 @@ class QEMUTracerBow(ContextBow):
 
             target_cmd = self._build_command(trace_filename=target_trace_filename, magic_filename=target_magic_filename, coredump_dir=tmpdir)
 
-            with self.target.run_context(target_cmd, timeout=self.timeout) as p:
-                r = QemuTraceResult()
-
-                try:
-                    yield Flight(self.target, p, r)
-                    r.timed_out = False
-                except subprocess.TimeoutExpired:
-                    r.timed_out = True
-
-            if not r.timed_out:
-                r.returncode = p.returncode
+            r = QemuTraceResult()
+            try:
+                with self.target.flight_context(target_cmd, timeout=self.timeout, result=r) as flight:
+                    yield flight
+            except subprocess.TimeoutExpired:
+                r.timed_out = True
+            else:
+                r.timed_out = False
+                r.returncode = flight.process.returncode
 
                 # did a crash occur?
                 if r.returncode in [ 139, -11 ]:
