+++ tests/test_bow_ltrace.py
@@ -1,6 +1,5 @@
 from time import sleep
 
-import nclib
 import archr
 import os
 
@@ -25,8 +24,8 @@ def ltrace_attach(t, p, **kwargs):
     b = archr.arsenal.LTraceAttachBow(t)
     pid = p.pid if isinstance(t, archr.targets.LocalTarget) else t.get_proc_pid('socat')
     with b.fire_context(pid=pid, ltrace_args=LTRACE_ARGS, **kwargs) as flight:
-        sleep(1)
-        nc = nclib.Netcat(flight.get_channel('tcp:0')) # misuse of flight
+        sleep(0.1)
+        nc = flight.get_channel('tcp:0') # misuse of flight
         nc.send(b'ahoi!')
         assert nc.readuntil(b'ahoi!', timeout=5) == b'ahoi!'
