+++ tests/test_bow_qemu.py
@@ -31,7 +31,7 @@ def crasher_checks(t):
     assert os.path.getsize(r.core_path) > 0
 
 def crash_on_input_checks(t):
-    crashing = "A"*120
+    crashing = b"A"*120
     b = archr.arsenal.QEMUTracerBow(t)
     r = b.fire(save_core=True, testcase=crashing)
 
@@ -76,8 +76,10 @@ def test_crasher_trace_local():
         crasher_checks(t)
 
 if __name__ == '__main__':
+    test_crash_on_input_trace()
     test_arrow_injection_docker()
     test_arrow_injection_local()
     test_crasher_trace()
     test_crasher_trace_local()
-    test_crash_on_input_trace()
+    test_vuln_stacksmash()
+    test_shellcode_tester()
